Overfitting detection: backtracking implemented.
Basically we save the best version of the network seen so far when there
are hints of possible overfitting ahead in the training process.

This was able to totally eliminate a few "bad runs" using AUTOSTOP with
the Iris dataset, where a sudden overfitting happened before the
autostop algorithm could fire.

This is an user option, since can slow down considerably things with
very large nets, however most of the times is the thing to do, so I'm
not sure if in the future should be the default.