Reverted: overfitting_limit set to 10 when backtrack is on.
In the practice with a limit of 10 the overfitting algorithm sometimes
does not kick in, making things too slow, even if sometimes this allows
to find a better fitting network.

It makes more sense for AUTOSTOP and BACKTRACK, to be two options that
don't aim at the best but aim to decent compromise between
generalization and training time for users that don't want to fine tune.