RDB encoding change: save deltas and past gradients to continue training
We don't just save weights but also deltas and p-gradient arrays in
order to be able to continue the training after loading the network from
the disk.

The previous encoding is no longer support, it was trivial to make it
back compatible but since the module is still beta and so new, better to
don't fill the code of what will soon be never used code.

NOTE: Now for the module to work, the latest Redis "unstable" commit is
needed. I had to add an API to the modules system.